Complete rewrite in terms of ytl
Originally, I thought that implementing a site wrapper required it to be
isomorphic to the base foundation site, because of some misunderstanding
of the types - this meant that all configuration had to be at the type
level, which was a big pain.

With some experience, I've now found that site wrappers can be done
much more easily, and the result is ytl. Using ytl, it's possible to
rewrite yesod-katip with a reader-based design, which allows for
configuration to live at the value level.

Some helpers from ytl have also simplified the code a lot. Hopefully
both libraries can be released to Hackage soon.